chore: move websearches to subcategory
